Fix static features tensoring

+++ deeplearning/clgen/active_models/data_generator.py
@@ -65,7 +65,7 @@ class DictPredictionDataloader(torch.utils.data.Dataset):
     for dp in dataset:
       self.dataset.append(
         {
-          'static_features' : dp['static_features'],
+          'static_features' : torch.FloatTensor(dp['static_features']),
           'input_ids'       : torch.FloatTensor(dp['input_ids']),
           'predictions'     : torch.LongTensor([-1, -1]),
         }
